move status render helpers to comp/core/status
